+++ README.md
@@ -1,4 +1,8 @@
-## TEST RUN INSTRUCTIONS
-* Start up a container with the image provided in the submission-methodology.json
-* Clone this git repository and navigate into the 0527291751/ folder
-* To test, execute python run.py test, and you should be able to see the outputs.
+# Spatially-explicit Agent Based Modeling for Evaluating School Reopening Plans
+
+As schools around San Diego county attempt to reopen amidst the COVID-19 pandemic, there is an increasing need to detect and quantify potentially risky activities in schools to help schools test out their individual reopening plans in order to prevent an outbreak. In this capstone project, we will use spatially explicit agent-based modeling techniques to help detect risky activities in individual schools by incorporating elements such as behavioral factors, environmental factors, and effects from non-pharmaceutical interventions (NPIs) in specific spatially constrained environments. You will have the opportunity to learn key concepts from spatial modeling, scalable computation, epidemiology, sociology, transportation and computational science from relevant domain experts. This is a unique opportunity to contribute to an immediate issue in our local community and helped them to drive impactful awareness using data science!
+
+---
+
+# Model Details
+Fundamental agent-based simulations are carried out using [mesa-geo](https://github.com/Corvince/mesa-geo) and open source spatial libraries such as [geopandas](https://geopandas.org/)
