+++ config/NPI.ini
@@ -1,7 +1,7 @@
-[NO_NPI]
-infection_distance = 18
-
-[MASKS]
-infection_distance = 6
+[MASK_PROB]
+cotton = 0.1
+multilayer = 0.2
+surgical = 0.6
+n95 = 0.1
