+++ config/activities.ini
@@ -1,7 +1,5 @@
 [CLASSROOM_REGULAR]
-seat_dist = 12
-infection_dist = 12
-viral_load_factor = 0.01
+seat_dist = 5
 move_spread = 12
 
 [LUNCH]
