+++ config/diseaseparams.ini
@@ -15,3 +15,4 @@ shape = 0.6432659248014824
 loc = -0.07787673726582335
 scale = 4.2489459496009125
 upper_bound = 17
+lower_bound = 0
