+++ config/schoolparams.ini
@@ -19,6 +19,8 @@ init_patient = 3
 mask_prob = 0.516
 population_attending = 1
 bus_prob = 0.04
+circular_desk_prob = 0.5
+
 
 [TIME]
 schedule_steps = 1
@@ -26,15 +28,15 @@ days = 1
 
 
 [INTERVENTION]
-mask_type = None
-ventilation_type = None
+ventilation_type = 'open_windows'
 teacher_vaccine_prob = 0.01
 student_vaccine_prob = 0
-teacher_testing_freq = w
-student_testing_freq = n
+teacher_testing_freq = 7
+student_testing_freq = 14
+
 
 [COHORT]
-radius = 8
+radius = 2
 size_prob = {1: 0.05, 2: 0.15, 3: 0.1, 4: 0.2, 5: 0.4, 6:0.1}
 same_room_prob = {0: 0.2, 1: 0.8}
 max_intersection_prob = 0.7
