--- notebooks/readme.MD
@@ -1,7 +1,0 @@
-# Please check copyscript_temporal
-do `ctrl+f` for the following <br>
-all updates are labeled as `UPDATE` <br>
-all naive work are labeled as `TODO` <br>
-
-currently the changes only exist in this notebook for testing purposes <br>
-once testing is done, code should be merged to `src/school_model.py`
