+++ schedule_data/day_schedule.csv
@@ -1,0 +1,91 @@
+schedule1,schedule2
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+recess,recess
+recess,recess
+recess,recess
+recess,class
+recess,class
+recess,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,lunch
+class,lunch
+class,lunch
+class,lunch
+class,lunch
+class,lunch
+lunch,lunch
+lunch,lunch
+lunch,recess
+lunch,recess
+lunch,recess
+lunch,recess
+lunch,recess
+lunch,recess
+recess,recess
+recess,recess
+recess,recess
+recess,recess
+recess,class
+recess,class
+recess,class
+recess,class
+recess,class
+recess,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+recess,recess
+recess,recess
+recess,recess
+class,recess
+class,recess
+class,recess
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
+class,class
