+++ schedule_data/small_schedule.csv
@@ -1,0 +1,6 @@
+schedule1,schedule2
+class,class
+class,class
+class,class
+class,class
+class,class
