+++ src/util.py
@@ -1,0 +1,34 @@
+from shapely.geometry import Point
+import random
+
+
+def generate_random(polygon):
+    minx, miny, maxx, maxy = polygon.bounds
+    while True:
+        pnt = Point(random.uniform(minx, maxx), random.uniform(miny, maxy))
+        if polygon.contains(pnt):
+            return pnt
+        
+        
+        
+        
+        
+        
+def find_room_type(room_agents, room_type):
+    """
+    room_agents: a list containing room agents
+    room_type: a valid string of room type: [None, 'restroom_grade_boys', 'lunch_room', 'classroom_grade',
+       'restroom_all', 'restroom_grade_girls', 'restroom_KG',
+       'classroom_KG', 'community_room', 'library',
+       'restroom_special_education', 'restroom_faculty',
+       'classroom_special_education', 'health_room', 'faculty_lounge',
+       'classroom_preschool', 'restroom_preschool']
+    """
+    out = []
+    for agent in room_agents:
+        if agent.room_type == room_type:
+            out.append(agent)
+    return out
+
+
+
