--- submission.json
@@ -1,9 +1,0 @@
-{
-    "config": ["config/activities.ini", "config/diseaseparams.ini",  "config/modelparams.ini",  "config/NPI.ini",  "config/schoolparams.ini",  "config/vizparams.ini"], 
-    "build": "run.py", 
-    "library": ["src"], 
-    "notebook": ["notebooks/run_notebook.ipynb"],
-    "targets": ["run-notebook", "analyze-results"],
-    "dockerhub-id": "jil1119/docker-mesa-geo:latest",
-    "repository-id": "https://github.com/MuscularBrain/0527291751.git"
-}
