+++ test/output/model_df.csv
@@ -1,6 +1,26 @@
 day,cov_positive
-0,0
-0,0
-0,0
-0,0
-0,0
+1,0
+1,0
+1,0
+1,0
+1,0
+2,0
+2,0
+2,0
+2,0
+2,0
+3,2
+3,2
+3,2
+3,2
+3,2
+4,10
+4,10
+4,10
+4,10
+4,10
+5,39
+5,39
+5,39
+5,39
+5,39
